move to biginteger

+++ data/model/user.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, String
 from config.dbConfig import Base
 
 
 class User(Base):
     __tablename__ = "users"
 
-    telegramId = Column(Integer, primary_key=True, index=True)
+    telegramId = Column(BigInteger, primary_key=True, index=True)
     userId = Column(String, index=True)
     accessToken = Column(String, index=True)
     refreshToken = Column(String, index=True)
