+++ data/model/verifiedLink.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, String
 
 from config.dbConfig import Base
 
@@ -7,7 +7,7 @@ class VerifiedLink(Base):
     __tablename__ = "verified_links"
 
     reference = Column(String, primary_key=True, index=True)
-    telegramId = Column(Integer, ForeignKey("users.telegramId"), nullable=False, index=True)
+    telegramId = Column(BigInteger, ForeignKey("users.telegramId"), nullable=False, index=True)
     userId = Column(String, nullable=False, index=True)
     name = Column(String, nullable=True)
     status = Column(String, nullable=True)
